+++ src/ui/panes/mod.rs
@@ -1,4 +1,4 @@
-use std::{borrow::Cow, cmp::Ordering, collections::HashMap};
+use std::{borrow::Cow, cmp::Ordering, collections::HashMap, time::Duration};
 
 use album_art::AlbumArtPane;
 use albums::AlbumsPane;
@@ -851,6 +851,24 @@ impl Property<PropertyKind> {
                         context.queue.len().with_thousands_separator(thousands_separator),
                         style,
                     )))
+                }
+                StatusProperty::QueueTimeTotal { separator } => {
+                    let sum: Duration = context.queue.iter().filter_map(|s| s.duration).sum();
+                    Some(Either::Left(Span::styled(sum.format_to_duration(separator), style)))
+                }
+                StatusProperty::QueueTimeRemaining { separator } => {
+                    let sum = context.find_current_song_in_queue().map_or(
+                        Duration::default(),
+                        |(current_song_idx, _)| {
+                            context
+                                .queue
+                                .iter()
+                                .skip(current_song_idx)
+                                .filter_map(|s| s.duration)
+                                .sum()
+                        },
+                    );
+                    Some(Either::Left(Span::styled(sum.format_to_duration(separator), style)))
                 }
             },
             PropertyKindOrText::Property(PropertyKind::Widget(w)) => match w {